+++ Source/Events.Store/EventLogSequenceNumber.cs
@@ -8,7 +8,7 @@ namespace Dolittle.Runtime.Events.Store
     /// <summary>
     /// Represents the sequence number of the Event Log as a natural number, corresponding to the number of events that has been committed to the Event Store.
     /// </summary>
-    public class EventLogSequenceNumber : ConceptAs<uint>
+    public class EventLogSequenceNumber : ConceptAs<ulong>
     {
         /// <summary>
         /// The initial sequence number of the Event Store before any Events are committed.
@@ -16,9 +16,9 @@ namespace Dolittle.Runtime.Events.Store
         public static EventLogSequenceNumber Initial = 0;
 
         /// <summary>
-        /// Implicitly convert a <see cref="uint"/> to an <see cref="EventLogSequenceNumber"/>.
+        /// Implicitly convert a <see cref="ulong"/> to an <see cref="EventLogSequenceNumber"/>.
         /// </summary>
         /// <param name="number">The number.</param>
-        public static implicit operator EventLogSequenceNumber(uint number) => new EventLogSequenceNumber { Value = number };
+        public static implicit operator EventLogSequenceNumber(ulong number) => new EventLogSequenceNumber { Value = number };
     }
 }
